requirements.txt
llama-index
google-generativeai
python-dotenv
chromadb
pandas
fastapi
uvicorn[standard]
crawl4ai
beautifulsoup4
requests
llama-index-embeddings-huggingface
sentence-transformers
llama-index-llms-gemini
llama-index-embeddings-gemini
llama-index-vector-stores-chroma
tavily-python
playwright
pytest
sqlalchemy
psycopg2-binary
nltk
